Move Firestore impl into persistence.remote.firestore
