+++ gnd/src/main/java/com/google/android/gnd/persistence/remote/firestore/DataStoreException.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.google.android.gnd.service.firestore;
+package com.google.android.gnd.persistence.remote.firestore;
 
 public class DataStoreException extends RuntimeException {
   public DataStoreException(String message) {
