+++ gnd/src/main/java/com/google/android/gnd/persistence/remote/firestore/DocumentNotFoundException.java
@@ -14,6 +14,6 @@
  * limitations under the License.
  */
 
-package com.google.android.gnd.service.firestore;
+package com.google.android.gnd.persistence.remote.firestore;
 
 public class DocumentNotFoundException extends Exception {}
