+++ gnd/src/main/java/com/google/android/gnd/persistence/remote/firestore/GndFirestore.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.google.android.gnd.service.firestore;
+package com.google.android.gnd.persistence.remote.firestore;
 
 import static java8.util.stream.Collectors.toList;
 import static java8.util.stream.StreamSupport.stream;
