+++ gnd/src/main/java/com/google/android/gnd/persistence/remote/firestore/UserDoc.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.google.android.gnd.service.firestore;
+package com.google.android.gnd.persistence.remote.firestore;
 
 import androidx.annotation.Nullable;
 import com.google.android.gnd.system.AuthenticationManager.User;
